Fix warning from compilers where sizeof(size_t) > sizeof(int)
Without this fix, a bunch of warnings are spewed

+++ radix_tree.hpp
@@ -34,7 +34,7 @@ int radix_length(const K &key);
 template<>
 inline int radix_length<std::string>(const std::string &key)
 {
-    return key.size();
+    return static_cast<int>(key.size());
 }
 
 template <typename K, typename T, typename Compare>
